README.md
This is a simple bluetooth-controlled car. The Micro-Controller (Arduino) is connected to the car.
For receiving data wirelessly we are using the HC-05 Bluetooth module.
At first, We have to connect/ pair the Bluetooth module with the phone (Android) you want to control.
F, B, L, R are used for moving the car Forward, Backward, Left, Right movements.